// src/api/gen/types/TokenDtoV1.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const tokenDtoV1VersionEnum = {
  V1: 'V1',
} as const

export type TokenDtoV1VersionEnumKey =
  (typeof tokenDtoV1VersionEnum)[keyof typeof tokenDtoV1VersionEnum]

export type TokenDtoV1 = {
  /**
   * @type string | undefined
   */
  version?: TokenDtoV1VersionEnumKey
  /**
   * @type string | undefined
   */
  address?: string
  /**
   * @type string | undefined
   */
  name?: string
  /**
   * @type string | undefined
   */
  smallestUnitName?: string
  /**
   * @type integer | undefined, int32
   */
  numberOfDecimals?: number
  /**
   * @type string | undefined
   */
  websiteUrl?: string
  /**
   * @type string | undefined
   */
  logoUrl?: string
  /**
   * @type string | undefined
   */
  maxSupply?: string
  /**
   * @type boolean | undefined
   */
  userBurnable?: boolean
  /**
   * @type string | undefined
   */
  originTxHash?: string
  /**
   * @type string | undefined
   */
  updatedByTxHash?: string
  /**
   * @type string | undefined
   */
  totalSupply?: string
  /**
   * @type integer | undefined, int64
   */
  createdAtBlockHeight?: number
  /**
   * @type integer | undefined, int64
   */
  updatedAtBlockHeight?: number
  /**
   * @type string | undefined, date-time
   */
  createdAtTimestamp?: string
  /**
   * @type string | undefined, date-time
   */
  updatedAtTimestamp?: string
}